--- feature_album/src/main/kotlin/com/igorwojda/showcase/feature/album/data/model/AlbumListDataModel.kt
@@ -1,5 +1,0 @@
-package com.igorwojda.showcase.feature.album.data.model
-
-internal data class AlbumListDataModel(
-    val album: List<AlbumDataModel>
-)
