--- feature_album/src/main/kotlin/com/igorwojda/showcase/feature/album/data/model/AlbumSearchResultDataModel.kt
@@ -1,7 +1,0 @@
-package com.igorwojda.showcase.feature.album.data.model
-
-import com.squareup.moshi.Json
-
-internal data class AlbumSearchResultDataModel(
-    @field:Json(name = "albummatches") val albumMatches: AlbumListDataModel
-)
