+++ feature_album/src/main/kotlin/com/igorwojda/showcase/feature/album/data/model/database/AlbumDataImageSizeDataDatabaseModel.kt
@@ -1,0 +1,10 @@
+package com.igorwojda.showcase.feature.album.data.model.database
+
+import com.igorwojda.showcase.feature.album.domain.model.AlbumImageDomainModel
+
+enum class AlbumDataImageSizeEntity {
+    MEDIUM, SMALL, LARGE, EXTRA_LARGE, MEGA;
+}
+
+internal fun AlbumImageDataEntity.toDomainModel() =
+    this.size.toDomainEnum()?.let { AlbumImageDomainModel(this.url, it) }
