+++ feature_album/src/main/kotlin/com/igorwojda/showcase/feature/album/data/model/database/AlbumImageDataDatabaseModel.kt
@@ -1,0 +1,8 @@
+package com.igorwojda.showcase.feature.album.data.model.database
+
+import com.igorwojda.showcase.feature.album.domain.enum.AlbumDomainImageSize
+
+data class AlbumImageDataEntity(val url: String, val size: AlbumDataImageSizeEntity)
+
+internal fun AlbumDataImageSizeEntity.toDomainEnum() =
+    AlbumDomainImageSize.values().firstOrNull { it.ordinal == this.ordinal }
