+++ feature_album/src/main/kotlin/com/igorwojda/showcase/feature/album/data/model/network/AlbumImageNetworkDataModel.kt
@@ -1,9 +1,9 @@
-package com.igorwojda.showcase.feature.album.data.model
+package com.igorwojda.showcase.feature.album.data.model.network
 
 import com.igorwojda.showcase.feature.album.data.enum.AlbumDataImageSize
 import com.igorwojda.showcase.feature.album.data.enum.toDomainEnum
 import com.igorwojda.showcase.feature.album.data.enum.toEntityEnum
-import com.igorwojda.showcase.feature.album.data.room.entity.AlbumImageDataEntity
+import com.igorwojda.showcase.feature.album.data.model.database.AlbumImageDataEntity
 import com.igorwojda.showcase.feature.album.domain.model.AlbumImageDomainModel
 import com.squareup.moshi.Json
