+++ feature_album/src/main/kotlin/com/igorwojda/showcase/feature/album/data/model/network/AlbumNetworkDataModel.kt
@@ -1,7 +1,7 @@
-package com.igorwojda.showcase.feature.album.data.model
+package com.igorwojda.showcase.feature.album.data.model.network
 
 import com.igorwojda.showcase.feature.album.data.enum.AlbumDataImageSize
-import com.igorwojda.showcase.feature.album.data.room.entity.AlbumDataEntity
+import com.igorwojda.showcase.feature.album.data.model.database.AlbumDataEntity
 import com.igorwojda.showcase.feature.album.domain.model.AlbumDomainModel
 import com.squareup.moshi.Json
