+++ feature_album/src/main/kotlin/com/igorwojda/showcase/feature/album/data/model/network/AlbumSearchResultNetworkDataModel.kt
@@ -1,0 +1,7 @@
+package com.igorwojda.showcase.feature.album.data.model.network
+
+import com.squareup.moshi.Json
+
+internal data class AlbumSearchResultNetworkDataModel(
+    @field:Json(name = "albummatches") val albumMatchesNetwork: AlbumListNetworkDataModel
+)
