+++ feature_album/src/main/kotlin/com/igorwojda/showcase/feature/album/data/retrofit/response/GetAlbumInfoResponse.kt
@@ -1,6 +1,6 @@
 package com.igorwojda.showcase.feature.album.data.retrofit.response
 
-import com.igorwojda.showcase.feature.album.data.model.AlbumDataModel
+import com.igorwojda.showcase.feature.album.data.model.network.AlbumDataModel
 
 internal data class GetAlbumInfoResponse(
     val album: AlbumDataModel
