+++ feature_album/src/test/kotlin/com/igorwojda/showcase/feature/album/data/model/AlbumDataModelTest.kt
@@ -2,6 +2,7 @@ package com.igorwojda.showcase.feature.album.data.model
 
 import com.igorwojda.showcase.feature.album.data.DataFixtures
 import com.igorwojda.showcase.feature.album.data.enum.AlbumDataImageSize
+import com.igorwojda.showcase.feature.album.data.model.network.toDomainModel
 import com.igorwojda.showcase.feature.album.domain.enum.AlbumDomainImageSize
 import com.igorwojda.showcase.feature.album.domain.model.AlbumDomainModel
 import org.amshove.kluent.shouldBeEqualTo
