+++ feature_album/src/test/kotlin/com/igorwojda/showcase/feature/album/data/model/AlbumWikiDataModelTest.kt
@@ -1,6 +1,7 @@
 package com.igorwojda.showcase.feature.album.data.model
 
 import com.igorwojda.showcase.feature.album.data.DataFixtures
+import com.igorwojda.showcase.feature.album.data.model.network.toDomainModel
 import com.igorwojda.showcase.feature.album.domain.model.AlbumWikiDomainModel
 import org.amshove.kluent.shouldBeEqualTo
 import org.junit.jupiter.api.Test
